+++ NEWS.md
@@ -9,6 +9,7 @@
 - fix time expansion bug in `timeAverage`
 - add new `type` "yearseason" (or "seasonyear"). This will split data by every year / season combination, making sure the seasons are contiguous. For example, in winter in teh northern hemisphere December 2010 will be considered part of winter 2011, rather than winter 2010. Thanks to Ralf Weisse for the suggestion.
 - fix download issues with `importAURN` and `importMeta` when users are within an organisational network. Problems likley due to move from http to https and SSL Certificates.
+- fix `TheilSen` bug when two types.
 
 # Changes in openair 1.8-2